--- AtsExCsTemplate.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33516.290
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VehiclePlugin", "VehiclePlugin\VehiclePlugin.csproj", "{B4C723A7-FD6A-4931-B3BF-A6635F44899B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MapPlugin", "MapPlugin\MapPlugin.csproj", "{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Extension", "Extension\Extension.csproj", "{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B4C723A7-FD6A-4931-B3BF-A6635F44899B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C7F17EA7-55CF-449F-9255-C3B1721A4DC9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{80D2BAF1-62D2-4538-8249-5F1C6D2B3DCE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7A01144D-3066-4B8F-9C06-A9212986B216}
-	EndGlobalSection
-EndGlobal
