+++ CommEx.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33516.290
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CommEx", "CommEx\CommEx.csproj", "{EE56A6F8-0E94-4E02-9EC9-E4E8CADAC3B0}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{EE56A6F8-0E94-4E02-9EC9-E4E8CADAC3B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE56A6F8-0E94-4E02-9EC9-E4E8CADAC3B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE56A6F8-0E94-4E02-9EC9-E4E8CADAC3B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE56A6F8-0E94-4E02-9EC9-E4E8CADAC3B0}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {7A01144D-3066-4B8F-9C06-A9212986B216}
+	EndGlobalSection
+EndGlobal
