+++ CommEx/CommEx.csproj
@@ -3,7 +3,7 @@
 
   <PropertyGroup>
     <TargetFramework>net48</TargetFramework>
-    <RootNamespace>AtsExCsTemplate.Extension</RootNamespace>
+    <RootNamespace>CommEx</RootNamespace>
     <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
     <Deterministic>false</Deterministic>
     <RestorePackagesWithLockFile>true</RestorePackagesWithLockFile>
