--- Extension/Extension.cs
@@ -1,46 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using AtsEx.PluginHost.Plugins;
-using AtsEx.PluginHost.Plugins.Extensions;
-
-namespace AtsExCsTemplate.Extension
-{
-    /// <summary>
-    /// プラグインの本体
-    /// Plugin() の第一引数でこのプラグインの仕様を指定
-    /// Plugin() の第二引数でこのプラグインが必要とするAtsEX本体の最低バージョンを指定（オプション）
-    /// </summary>
-    [Plugin(PluginType.Extension)]
-    internal class ExtensionMain : AssemblyPluginBase, IExtension
-    {
-        /// <summary>
-        /// プラグインが読み込まれた時に呼ばれる
-        /// 初期化を実装する
-        /// </summary>
-        /// <param name="builder"></param>
-        public ExtensionMain(PluginBuilder builder) : base(builder)
-        {
-        }
-
-        /// <summary>
-        /// プラグインが解放されたときに呼ばれる
-        /// 後処理を実装する
-        /// </summary>
-        public override void Dispose()
-        {
-        }
-
-        /// <summary>
-        /// シナリオ読み込み中に毎フレーム呼び出される
-        /// </summary>
-        /// <param name="elapsed">前回フレームからの経過時間</param>
-        public override TickResult Tick(TimeSpan elapsed)
-        {
-            return new ExtensionTickResult();
-        }
-    }
-}
